Better Error Handling (#11)
* Added Appropriate PHPCS Error

When the worker receives an error from PHPCS we should
take care to handle it in a special way to inform the user
of the error. This commit introduces a dedicated PHPCSError
type for this purpose and rejects the promise with it.

* Added PHPCS Error Presentation

A new logger class will allow us to centralize logging messages
into an output channel and settle on a consistent place to
handle errors and eventually other logging messages.

* Added Update Dirty Checking

Rather than updating as often as events arrive we should be
checking the version. This avoids unnecessarily executing
workers such as when an autosave comes after a change.

* Changelog