Clear Status When Document Closes (#30)
When a diagnostic is cancelled, we were leaving behind the status
entry to dangle. This adds a cancellation event to stop the status,
thus removing it from the status bar when cancelled. We were also
using the Uri as a key in many different places, but this led to bugs
when they were not identical objects (even though identical Uris).
For convenience, we've wrapped these places with a class that
gets the string from the Uri as the key.