Restructured PHPCS Integration
Since PHPCS can't reliably detect the FQN of the report class, I've
restructured the PHPCS integration so that it will ensure only a
single class file is ever loaded when including the report.
This will ensure we always load the correct report file.

This commit also comes with some changes to the build process.
We don't need to reference the report files anymore, since we aren't
going to require Webpack to compile the assets to find the reports.