Stabilize Working Directory (#65)
Changing the working directory based on whether or not an executable
was automatically found can lead to inconsistent behavior when using
relative paths with `phpCodeSniffer.standardCustom`. Initially this was
done so that the `"Default"` standard option would let PHPCS traverse
up to the workspace root. Since the `"Automatic"` standard type finds
the nearest PHPCS configuration, this is not necessary.