Support Overriding PHPCS Integration Path (#89)
* Added Special Options Configuration

There are currently some very specific use-cases that
we don't support. In order to add features like
that without bloating our config we can use
an aggregate config setting that only gets
filled out when a user needs to set
one of the parameters.

The first of these is an option to override the path to
the extension's PHPCS integration files.